rename: Chip component -> OptionChip
